feat(collector): relocation collector main type from model
