+++ apps/web/src/services/asset-inventory/collector/collector-main/modules/CollectorContents.vue
@@ -85,7 +85,7 @@ import CollectorScheduleModal
 import {
     COLLECTOR_ITEM_INFO_TYPE,
     COLLECTOR_QUERY_HELPER_SET,
-} from '@/services/asset-inventory/collector/model';
+} from '@/services/asset-inventory/collector/collector-main/type';
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/route-config';
 
 interface Props {
